requirements.txt
Django==5.2.9
djangorestframework==3.16.1
python-decouple==3.8
requests==2.32.5
Pillow==12.0.0
psycopg2-binary==2.9.11
google-cloud-vision>=3.0.0,<4.0.0
google-generativeai
